+++ kubernetes-model-generator/openshift-model-machineconfig/cmd/generate/generate.go
@@ -1,0 +1,105 @@
+/**
+ * Copyright (C) 2015 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package main
+
+import (
+  "bytes"
+  "encoding/json"
+  "fmt"
+  metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+  apimachineryversion "k8s.io/apimachinery/pkg/version"
+  runtime "k8s.io/apimachinery/pkg/runtime"
+  "log"
+  "reflect"
+  "strings"
+  "time"
+  machineconfig "github.com/openshift/machine-config-operator/pkg/apis/machineconfiguration.openshift.io/v1"
+
+  "os"
+
+  "github.com/fabric8io/kubernetes-client/kubernetes-model-generator/pkg/schemagen"
+)
+
+type Schema struct {
+  Info                                     apimachineryversion.Info
+  APIGroup                                 metav1.APIGroup
+  APIGroupList                             metav1.APIGroupList
+  BaseKubernetesList                       metav1.List
+  ObjectMeta                               metav1.ObjectMeta
+  TypeMeta                                 metav1.TypeMeta
+  Status                                   metav1.Status
+  Patch                                    metav1.Patch
+  Time                                     metav1.Time
+  ContainerRuntimeConfig                   machineconfig.ContainerRuntimeConfig
+  ContainerRuntimeConfigList               machineconfig.ContainerRuntimeConfigList
+  ControllerConfig                         machineconfig.ControllerConfig
+  ControllerConfigList                     machineconfig.ControllerConfigList
+  KubeletConfig                            machineconfig.KubeletConfig
+  KubeletConfigList                        machineconfig.KubeletConfigList
+  MachineConfigPool                        machineconfig.MachineConfigPool
+  MachineConfigPoolList                    machineconfig.MachineConfigPoolList
+  MachineConfig                            machineconfig.MachineConfig
+  MachineConfigList                        machineconfig.MachineConfigList
+}
+
+func main() {
+  packages := []schemagen.PackageDescriptor{
+    {"k8s.io/api/core/v1", "", "io.fabric8.kubernetes.api.model", "kubernetes_core_", false},
+    {"k8s.io/apimachinery/pkg/api/resource", "", "io.fabric8.kubernetes.api.model", "kubernetes_resource_", false},
+    {"k8s.io/apimachinery/pkg/util/intstr", "", "io.fabric8.kubernetes.api.model", "kubernetes_apimachinery_pkg_util_intstr_", false},
+    {"k8s.io/apimachinery/pkg/runtime", "", "io.fabric8.openshift.api.model.runtime", "kubernetes_apimachinery_pkg_runtime_", false},
+    {"k8s.io/apimachinery/pkg/version", "", "io.fabric8.kubernetes.api.model.version", "kubernetes_apimachinery_pkg_version_", false},
+    {"k8s.io/kubernetes/pkg/util", "", "io.fabric8.kubernetes.api.model", "kubernetes_util_", false},
+    {"k8s.io/kubernetes/pkg/api/errors", "", "io.fabric8.kubernetes.api.model", "kubernetes_errors_", false},
+    {"k8s.io/kubernetes/pkg/api/unversioned", "", "io.fabric8.kubernetes.api.model", "api_", false},
+    {"k8s.io/apimachinery/pkg/apis/meta/v1", "", "io.fabric8.kubernetes.api.model", "kubernetes_apimachinery_", false},
+    {"github.com/openshift/api/config/v1", "", "io.fabric8.openshift.api.model", "os_config_v1_", false},
+    {"github.com/openshift/machine-config-operator/pkg/apis/machineconfiguration.openshift.io/v1", "machineconfiguration", "io.fabric8.openshift.api.model.machineconfig.v1", "os_machineconfig_v1_", true},
+  }
+
+  typeMap := map[reflect.Type]reflect.Type{
+    reflect.TypeOf(time.Time{}): reflect.TypeOf(""),
+    reflect.TypeOf(struct{}{}):  reflect.TypeOf(""),
+  }
+  manualTypeMap := map[reflect.Type]string {
+    reflect.TypeOf(runtime.RawExtension{}): "java.util.Map<String, Object>",
+  }
+  schema, err := schemagen.GenerateSchema(reflect.TypeOf(Schema{}), packages, typeMap, manualTypeMap,"machineconfig")
+  if err != nil {
+    fmt.Fprintf(os.Stderr, "An error occurred: %v", err)
+    return
+  }
+
+  args := os.Args[1:]
+  if len(args) < 1 || args[0] != "validation" {
+    schema.Resources = nil
+  }
+
+  b, err := json.Marshal(&schema)
+  if err != nil {
+    log.Fatal(err)
+  }
+  result := string(b)
+  result = strings.Replace(result, "\"additionalProperty\":", "\"additionalProperties\":", -1)
+
+  var out bytes.Buffer
+  err = json.Indent(&out, []byte(result), "", "  ")
+  if err != nil {
+    log.Fatal(err)
+  }
+
+  fmt.Println(out.String())
+}