+++ openshift-client/src/main/java/io/fabric8/openshift/client/CodeGen.java
@@ -31,6 +31,7 @@ import io.sundr.transform.annotations.VelocityTransformations;
                 @ResourceSelector("operatorhub.properties"),
                 @ResourceSelector("monitoring.properties"),
                 @ResourceSelector("console.properties"),
+                @ResourceSelector("machineconfig.properties"),
         }
 
 )